Update init script, single-site, default, rm bin;
There isn't a need for ./bin to be available because we can just
leverage wp-cli for downloading WordPress. There is now a single-site
directory which will allow for different provisioning scripts for
different types of sites (i.e: WordPress Multisites). The readme is
updated to show that new differentiation.

The default project name is now project instead of the convoluted
project-name.