Update readme, split & add multisite configs;
The readme is now more robust and explains a lot of what needs to be done to use the boilerplate. The multi-site directory has now also been split into subdomains and subdirectories prefixes. They mostly do the same thing, but a boilerplate should be drop-in-and-go so there are two similar versions.